Update hcmt.py
Improved - dependency installation for macOS

Added - Server Specs population is now based on the Server Location you select.  It will only show you Server Specs available in that selected location. It also filters out ARM chip architecture so that it doesn't show them. Since ARM chip support is not available in nodectl at this time.

Added - The default Server Specs first tries to implement cx52 as the default if it is available in the chosen location. If it is not available it tries to set cpx51 as the default if it is available in the chosen location.  If neither are available, then no default is given.